Move `handle_gossip_message` into gossip engine
