+++ crates/sc-proof-of-time/src/node_client.rs
@@ -1,24 +1,19 @@
 //! Consensus node interface to the time keeper network.
 
 use crate::gossip::PotGossip;
-use crate::state_manager::PotProtocolState;
 use crate::PotComponents;
 use futures::StreamExt;
 use sc_client_api::BlockchainEvents;
-use sc_network::PeerId;
 use sc_network_gossip::{Network as GossipNetwork, Syncing as GossipSyncing};
 use sp_blockchain::HeaderBackend;
 use sp_consensus_subspace::digests::extract_pre_digest;
 use sp_core::H256;
 use sp_runtime::traits::Block as BlockT;
 use std::sync::Arc;
-use std::time::Instant;
-use subspace_core_primitives::PotProof;
 use tracing::{debug, error, trace, warn};
 
 /// The PoT client implementation
 pub struct PotClient<Block: BlockT<Hash = H256>, Client> {
-    pot_state: Arc<dyn PotProtocolState>,
     gossip: PotGossip<Block>,
     client: Arc<Client>,
 }
@@ -40,7 +35,6 @@ where
         GossipSync: GossipSyncing<Block> + 'static,
     {
         Self {
-            pot_state: components.protocol_state.clone(),
             gossip: PotGossip::new(
                 network,
                 sync,
@@ -54,13 +48,7 @@ where
     /// Runs the node client processing loop.
     pub async fn run(self) {
         self.initialize().await;
-        let handle_gossip_message: Arc<dyn Fn(PeerId, PotProof) + Send + Sync> =
-            Arc::new(|sender, proof| {
-                self.handle_gossip_message(sender, proof);
-            });
-        self.gossip
-            .process_incoming_messages(handle_gossip_message)
-            .await;
+        self.gossip.process_incoming_messages().await;
         error!("Gossip engine has terminated");
     }
 
@@ -105,19 +93,6 @@ where
                 );
                 return;
             }
-        }
-    }
-
-    /// Handles the incoming gossip message.
-    fn handle_gossip_message(&self, sender: PeerId, proof: PotProof) {
-        let start_ts = Instant::now();
-        let ret = self.pot_state.on_proof_from_peer(sender, &proof);
-        let elapsed = start_ts.elapsed();
-
-        if let Err(error) = ret {
-            trace!(%error, %sender, "On gossip");
-        } else {
-            trace!(%proof, ?elapsed, %sender, "On gossip");
         }
     }
 }
